+++ CHANGELOG.md
@@ -17,6 +17,7 @@ Note that since we don't clearly distinguish between a public and private interf
 - MolViewSpec: MVP Support for geometrical primitives (mesh, lines, line, label, distance measurement)
 - Mesoscale Explorer: Add support for 4-character PDB IDs (e.g., 8ZZC) in PDB-Dev loader
 - Fix Sequence View in Safari 18
+- Improve performance of `IndexPairBonds` assignment when operator keys are available
 
 ## [v4.7.1] - 2024-09-30